+++ regex/src/com.oracle.truffle.regex/src/com/oracle/truffle/regex/tregex/automaton/SmallStateSetImpl.java
@@ -90,7 +90,7 @@ final class SmallStateSetImpl<S> implements StateSet<S> {
         return toBit(stateIndex.getId((S) o));
     }
 
-    private static long toBit(short id) {
+    private static long toBit(int id) {
         assert 0 <= id && id < 64;
         return 1L << id;
     }
@@ -108,7 +108,7 @@ final class SmallStateSetImpl<S> implements StateSet<S> {
         return notPresent;
     }
 
-    private void removeId(short id) {
+    private void removeId(int id) {
         assert (set & toBit(id)) != 0;
         set &= ~toBit(id);
     }
@@ -260,7 +260,7 @@ final class SmallStateSetImpl<S> implements StateSet<S> {
         @Override
         public void remove() {
             assert bitIndex > 0;
-            stateSet.removeId((short) (bitIndex - 1));
+            stateSet.removeId(bitIndex - 1);
         }
     }
 }
