+++ regex/src/com.oracle.truffle.regex/src/com/oracle/truffle/regex/tregex/nfa/NFA.java
@@ -220,7 +220,7 @@ public final class NFA implements StateIndex<NFAState>, JsonConvertible {
     }
 
     @Override
-    public short getId(NFAState state) {
+    public int getId(NFAState state) {
         return state.getId();
     }
